requirements/requirements_t3.txt
javalang==0.13.0
requests==2.32.5 
tqdm>=4.0.0
python-dotenv>=1.2.1
pyyaml>=6.0.3
colorama
codereader
